basic functions done

+++ project-management-system-backend/Models/Budget.cs
@@ -4,12 +4,12 @@
     {
         public int Id { get; set; }
         public int ProjectId { get; set; }
-        public int TotalBudget { get; set; }  
-        public int HourlyCost { get; set; }
-        public int Actualcost { get; set; }
-        public int Plannedcost { get; set; }
-        public int Received { get; set; }
-        public int yetToReceive { get; set; }
+        public double TotalBudget { get; set; }  
+        public double HourlyCost { get; set; }
+        public double Actualcost { get; set; }
+        public double Plannedcost { get; set; }
+        public double Received { get; set; }
+        public double yetToReceive { get; set; }
